fixe the problem of hashing
