+++ app.js
@@ -19,79 +19,10 @@ const override = require('method-override')
 const session = require('express-session')
 const reg = require('./routes/register')
 const db = require('./bd')
-
-
 require('dotenv').config()				
 
-//Init
 const app = express()
-initializePassport(
-	passport,
-	(email) =>
-		reg.users.find((user) => {
-			return user.email === email
-		}),
-	(id) => reg.users.find((user) => user.id === id)
-)
 
-
-db.query("SELECT `id_u`, `email`, `motdepasse`, `civilite`, `prenom`, `nom`, `datenaissance`, `telephone`, `photo`, `admin`, `dateinscription` FROM `utilisateur` WHERE 1", function (err, result) {
-	if (err) throw err;
-	result.map(u => {
-		const user = {
-			id: u.id_u,
-			email: u.email,
-			password: u.motdepasse,
-			civility: u.civilite,
-			firstName: u.prenom,
-			lastName: u.nom,
-			birthDate: u.datenaissance,
-			phone: u.telephone,
-		}
-		reg.users.push(user)
-	})
-	
-});
-
-
-
-
-// db.connect(function(err) {
-//     if (err) throw err;
-    // db.query("SELECT `id_u`, `email`, `motdepasse`, `civilite`, `prenom`, `nom`, `datenaissance`, `telephone`, `photo`, `admin`, `dateinscription` FROM `utilisateur` WHERE 1", function (err, result) {
-    //     if (err) throw err;
-    //     result.map(u => {
-	// 		const user = {
-	// 			id: u.id_u,
-	// 			email: u.email,
-	// 			password: u.motdepasse,
-	// 			civility: u.civilite,
-	// 			firstName: u.prenom,
-	// 			lastName: u.nom,
-	// 			birthDate: u.datenaissance,
-	// 			phone: u.telephone,
-	// 		}
-    //         reg.users.push(u)
-    //     })
-    //     console.log(reg.users);
-    // });
-//     console.log("Connecté à la base de données MySQL!");
-// });
-
-//Database
-// const connection = mysql.createConnection({
-// 	host: 'localhost',
-// 	user: 'root',
-// 	password: '',
-// 	database: 'midrent',
-// })
-
-
-
-// connection.connect((err) => {
-// 	if (err) throw err
-// 	console.log('connected')
-// })
 
 //Middlewares
 app.use('/css', express.static(path.join(__dirname, 'public/css')))
@@ -121,6 +52,7 @@ app.use(passport.initialize())
 app.use(passport.session())
 app.use(override('_method'))
 
+
 //Routes
 app.use('/', require('./routes/index'))
 app.use('/login', require('./routes/login'))
@@ -130,16 +62,12 @@ app.use('/pub', require('./routes/pub'))
 app.use('/locals', require('./routes/locals'))
 app.use('/dashboard', require('./routes/dashboard'))
 app.use('/logout', require('./routes/logout'))
-app.post(
-	'/login',
-	passport.authenticate('local', {
-		successRedirect: '/',
-		failureRedirect: '/login',
-		failureFlash: true,
-	})
-)
+
+
+
 //View engine
 app.set('view engine', 'ejs')
+
 
 //Listener
 const port = process.env.PORT
